% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piececor-package.R
\docType{package}
\name{piececor-package}
\alias{piececor}
\alias{piececor-package}
\title{piececor: Tidy Piecewise Correlations}
\description{
To learn more about piececor see https://github.com/brshallo/piececor
}
\section{Notes}{

\itemize{
\item After making more flexible change \code{mgcv} to suggests in DESCRIPTION
}
}

\author{
\strong{Maintainer}: Bryan Shalloway \email{brshallodev@gmail.com}

}
\keyword{internal}
