% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/piecewise_cors.R
\name{piecewise_cors}
\alias{piecewise_cors}
\title{Piecewise Correlations}
\usage{
piecewise_cors(
  data,
  .target,
  ...,
  custom_model_spec = NULL,
  fit_formula = ".target ~ s(...)",
  cor_function = "~cor(.x, method = 'spearman')[2,1]"
)
}
\arguments{
\item{data}{Dataframe containing columns of interest.}

\item{.target}{Column name of target (unquoted). Numeric with
sufficient unique observations.}

\item{...}{Column names of variables of interest (unquoted), or a tidy
selection specification (see:
https://dplyr.tidyverse.org/reference/select.html). Separate models are
created for each variable passed into \code{...}. Numeric with sufficient unique
observations.}

\item{custom_model_spec}{\code{NULL} or model specification from the \code{{parsnip}}
package. Default is \code{NULL} in which case the model specification will be:\preformatted{parsnip::gen_additive_mod() \%>\%
  parsnip::set_engine("mgcv", method = "REML") \%>\%
  parsnip::set_mode("regression")
}

(At this point) must be a generalized additive model with \code{{mgcv}} engine.
See for details on parameters:
https://parsnip.tidymodels.org/reference/details_gen_additive_mod_mgcv.html}

\item{fit_formula}{Character string. Pseudo representation of formula passed
to \code{parsnip::fit()} of model specification. Default is \code{".target ~ s(...)"}. Note that \code{.target} and \code{...} must be included in any custom
\code{fit_formula}.}

\item{cor_function}{Character string of a lambda function passed to
\verb{purrr::map*()} to calculate correlation between \code{.target} and variables of
interest at piecewise segments of observations. Default is:
\code{"~cor(.x, method = 'spearman')[2,1]"}

Should be written in tidyverse friendly shortcut lambda notation (like
above), as opposed to traditional lambda function notation: \code{"function(x) cor(x, method = 'spearman')[2,1]"}

Must evaluate to a numeric vector of length one. Could pass in functions
specification unrelated to correlation.}
}
\value{
Named list of named list of models \verb{$mods} and named list of
dataframes \verb{$cors} for each variable of interest specified in \code{...}. Each
dataframe in \code{.cors} has the columns:

\code{gtoe}: (greater than or equal to)
\code{lt}: (less than)
\code{data}: list column containing dataframes of observations within segment
\code{n_obs}: number of rows in \code{data} segment
\code{cor}: value returned by \code{cor_function} for segment of observations
}
\description{
Independently for each variable of interest \code{...} and \code{.target}, calculate
the piecewise correlation. Observations included for each correlation
calculation are segmented based on local extrema (minima / maxima) of the
smoothed model fit.
}
\details{
If the smoothing model has no local minima / maxima, correlation is
calculated across all observations.
}
\examples{
# See README
}
