% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_splits.R
\name{plot_splits}
\alias{plot_splits}
\title{Plot Splits and Smoothed Model Fit}
\usage{
plot_splits(cors_mods, ..., show_points = TRUE)
}
\arguments{
\item{cors_mods}{List of lists of \code{mods} and \code{cors} of the form returned by
running \code{piececor::piecewise_cors()}.}

\item{...}{Column names of variables of interest (unquoted), or a tidy
selection specification (see:
https://dplyr.tidyverse.org/reference/select.html). Only specify for those
results you \emph{wish to plot}.}

\item{show_points}{Logical. Default is \code{TRUE}. Whether to include points in
the graph or not. If you have lots of observations, you may want to set to
\code{FALSE}.}
}
\value{
Plots of smoothing functions and splits between segments of
observations (vertical lines) on observations for each variable specified
in \code{...} (on x axis, \code{.target} variable will be on y).
}
\description{
Plot data with cut points for segments and smoothed model overlaid on
observations.
}
\examples{
# See README
}
