% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/weighted_abs_mean_cors.R
\name{weighted_abs_mean_cors}
\alias{weighted_abs_mean_cors}
\title{Weighted Absolute Mean of Correlations}
\usage{
weighted_abs_mean_cors(cors_mods, fisher_z_transform = TRUE)
}
\arguments{
\item{cors_mods}{List of lists of \code{mods} and \code{cors} of the form returned by running
\code{piececor::piecewise_cors()}.}

\item{fisher_z_transform}{Logical. Default is \code{TRUE}. If \code{FALSE}, simply
calculates the means of \code{cor} weighted by \code{n_obs} for each dataframe in
\code{cors}.

If \code{TRUE} applies Fisher's z transformation on correlations of
individual segments prior to calculating weighted mean (then applies
inverse).}
}
\value{
Dataframe with variables of interest in \code{name} column and weighted
mean correlation (across segments) in \code{value} for each dataframe in \code{cors_mods$cors}.
}
\description{
Returns the weighted mean of the absolute value of the correlations (with
Fisher's Z transformation applied if \code{fisher_z_transform} is left as \code{TRUE}).
}
\section{Warning}{
 See https://en.wikipedia.org/wiki/Fisher_transformation and
https://www.researchgate.net/post/How-to-calculate-the-weighted-average-of-correlations-per-outcome
for information on Fisher's z transformation. Appropriate methods for
calculating weighted correlations should be further investigated.
}

\examples{
# See README
}
